// packages/record-creator/src/types.ts
export interface RecordItemModel {
    name: string;
    checked: boolean;
}